+++ UnitTests/Collections/ObservableGroupTests.cs
@@ -2,36 +2,44 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.ObjectModel;
-using System.Collections.Specialized;
 using System.Linq;
 using FluentAssertions;
-using Microsoft.Toolkit.Observables.Collections;
+using Microsoft.Toolkit.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Collections.Specialized;
 
-namespace UnitTests.Observables
+namespace UnitTests.Collections
 {
     [TestClass]
-    public class ReadOnlyObservableGroupTests
+    public class ObservableGroupTests
     {
         [TestCategory("Observables")]
         [TestMethod]
-        public void Ctor_WithKeyAndCollection_ShouldHaveExpectedInitialState()
+        public void Ctor_ShouldHaveExpectedState()
         {
-            var source = new ObservableCollection<int>(new[] { 1, 2, 3 });
-            var group = new ReadOnlyObservableGroup<string, int>("key", source);
+            var group = new ObservableGroup<string, int>("key");
 
             group.Key.Should().Be("key");
+            group.Should().BeEmpty();
+        }
+
+        [TestCategory("Observables")]
+        [TestMethod]
+        public void Ctor_WithGrouping_ShouldHaveExpectedState()
+        {
+            var source = new IntGroup("Key", new[] { 1, 2, 3 });
+            var group = new ObservableGroup<string, int>(source);
+
+            group.Key.Should().Be("Key");
             group.Should().BeEquivalentTo(new[] { 1, 2, 3 }, option => option.WithStrictOrdering());
         }
 
         [TestCategory("Observables")]
         [TestMethod]
-        public void Ctor_ObservableGroup_ShouldHaveExpectedInitialState()
+        public void Ctor_WithCollection_ShouldHaveExpectedState()
         {
             var source = new[] { 1, 2, 3 };
-            var sourceGroup = new ObservableGroup<string, int>("key", source);
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", source);
 
             group.Key.Should().Be("key");
             group.Should().BeEquivalentTo(new[] { 1, 2, 3 }, option => option.WithStrictOrdering());
@@ -43,11 +51,10 @@ namespace UnitTests.Observables
         {
             var collectionChangedEventRaised = false;
             var source = new[] { 1, 2, 3 };
-            var sourceGroup = new ObservableGroup<string, int>("key", source);
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", source);
             ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-            sourceGroup.Add(4);
+            group.Add(4);
 
             group.Key.Should().Be("key");
             group.Should().BeEquivalentTo(new[] { 1, 2, 3, 4 }, option => option.WithStrictOrdering());
@@ -60,11 +67,10 @@ namespace UnitTests.Observables
         {
             var collectionChangedEventRaised = false;
             var source = new[] { 1, 2, 3 };
-            var sourceGroup = new ObservableGroup<string, int>("key", source);
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", source);
             ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-            sourceGroup[1] = 4;
+            group[1] = 4;
 
             group.Key.Should().Be("key");
             group.Should().BeEquivalentTo(new[] { 1, 4, 3 }, option => option.WithStrictOrdering());
@@ -77,11 +83,10 @@ namespace UnitTests.Observables
         {
             var collectionChangedEventRaised = false;
             var source = new[] { 1, 2, 3 };
-            var sourceGroup = new ObservableGroup<string, int>("key", source);
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", source);
             ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-            sourceGroup.Remove(1);
+            group.Remove(1);
 
             group.Key.Should().Be("key");
             group.Should().BeEquivalentTo(new[] { 2, 3 }, option => option.WithStrictOrdering());
@@ -94,11 +99,10 @@ namespace UnitTests.Observables
         {
             var collectionChangedEventRaised = false;
             var source = new[] { 1, 2, 3 };
-            var sourceGroup = new ObservableGroup<string, int>("key", source);
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", source);
             ((INotifyCollectionChanged)group).CollectionChanged += (s, e) => collectionChangedEventRaised = true;
 
-            sourceGroup.Clear();
+            group.Clear();
 
             group.Key.Should().Be("key");
             group.Should().BeEmpty();
@@ -111,8 +115,7 @@ namespace UnitTests.Observables
         [DataRow(3)]
         public void IReadOnlyObservableGroup_ShouldReturnExpectedValues(int count)
         {
-            var sourceGroup = new ObservableGroup<string, int>("key", Enumerable.Range(0, count));
-            var group = new ReadOnlyObservableGroup<string, int>(sourceGroup);
+            var group = new ObservableGroup<string, int>("key", Enumerable.Range(0, count));
             var iReadOnlyObservableGroup = (IReadOnlyObservableGroup)group;
 
             iReadOnlyObservableGroup.Key.Should().Be("key");
