+++ UnitTests/Collections/ObservableGroupedCollectionTests.cs
@@ -3,12 +3,12 @@
 // See the LICENSE file in the project root for more information.
 
 using FluentAssertions;
-using Microsoft.Toolkit.Observables.Collections;
+using Microsoft.Toolkit.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace UnitTests.Observables
+namespace UnitTests.Collections
 {
     [TestClass]
     public class ObservableGroupedCollectionTests
