+++ UnitTests/UnitTests.csproj
@@ -161,11 +161,11 @@
     <Compile Include="Markdown\Parse\SuperscriptTests.cs" />
     <Compile Include="Markdown\Parse\TableTests.cs" />
     <Compile Include="Markdown\TestBase.cs" />
-    <Compile Include="Observables\IntGroup.cs" />
-    <Compile Include="Observables\ObservableGroupedCollectionTests.cs" />
-    <Compile Include="Observables\ObservableGroupTests.cs" />
-    <Compile Include="Observables\ReadOnlyObservableGroupedCollectionTests.cs" />
-    <Compile Include="Observables\ReadOnlyObservableGroupTests.cs" />
+    <Compile Include="Collections\IntGroup.cs" />
+    <Compile Include="Collections\ObservableGroupedCollectionTests.cs" />
+    <Compile Include="Collections\ObservableGroupTests.cs" />
+    <Compile Include="Collections\ReadOnlyObservableGroupedCollectionTests.cs" />
+    <Compile Include="Collections\ReadOnlyObservableGroupTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Helpers\Test_WeakEventListener.cs" />
     <Compile Include="UI\Controls\Test_UniformGrid_AutoLayout.cs" />
